[base] Add a lightweight function wrapper - function_ref
This class is supposed to land in C++23, and allows for a much
more performant passing of generic callbacks. It's a non-owning
view of a function, similar to what string_view is to string.
What this means is one has to be careful when using it to never
create a dangling reference.

The best use case is passing a callback into some function that
is going to use it and never stores it for later.

Bug: 180535478
Test: atest libbase_test

Change-Id: I9ddd31cc6bcd012bf157f5f8215802426593530c