Add OR_RETURN and OR_FATAL macros
OR_RETURN is a macro that, evaluates a given statement, and then either
yields a success value or returns from the surrounding function with an
error value. It is generic so that it accepts any type of statement and
can be used in functions of any return type.

OR_FATAL is similar to OR_RETURN, but aborts when the statement yields
an error value.

For example:

\#include <android-base/result.h>

Result<T, E> inner();

Result<U, E> outer() {
  // if inner() is a success, the success value is set to t.
  // if inner() is a failure, the error of type ResultError<E> is
  // returned to the caller.
  T t = OR_RETURN(inner());
  ...
}

E anotherOuter() {
  // The same macro works in a diffrent function returning E
  // if inner() is a failure, the error of type E is returned to the
  // caller.
  T t = OR_RETURN(inner());
  ...
}

Internally, it relies on a template heler class OkOrFail<T>.
For a statement whose type is T, the specialization of the class for T
has to be provided. The specialization is responsible for providing
functions that the OR_RETURN macro uses. e.g. the member function
bool is_ok(const T&) provides a way to test if T is a success of not.

In this CL, the specialization for T = Result<V, E> is provided.

Bug: 209929099
Test: atest libbase_test
Change-Id: I66c95e2ebcd063577c81f0facbfe1cec60818a21